Rename file video.vue to Video.vue
